New bestpracticeprofile element to replace the lists element
Replace the very old list/sp-list/sp-view system, which was a visual
representation of how standards was selected for NATO common funded
systems.

The new system is an improvement from a semantic point of view, but still
a quisk and dirty last minute hack befor the release of NISP v9.

* The lists element are directly mapped to the bestpracticeprofile
  element
* The sp-list elements are directly mapped to bpserviceprofile elements
* The sp-view and the select elements are replaced with
  bpgroup are bprefstandard - but there are not a one-to-one mapping between
  the elements.