Implement an incomplete C3 Taxonomy v2 with new ID system
Implement a incomplete version 2.0 of the C3 Taxonomy using a new ID system. The ID system is replaced with a new system based on UUID's from the C3 taxonomy. Each ID is generated automatically with a T- prefix and a -X suffix.
There are currently 21 nodes, which are not included in a taxonomy due to a flawed data export from the EM-Wiki.