Add profilespec and status element to all profile elements.
This profilespec is similar to the document element used in standards, and describes the document describing a profile.

The status element is identical to the status element used in a standard.

Both element are right now optional, but will be mandatory as soon as we have filled out the data. We will initially only add these element to the capability profile, as the profile and service profile will automatically inherit the profilespec and status of its ancestor capability profile element.