Automate using the mode attribute on the status element
The mode attribute on status have never been used before as an indicator, if a standard was deleted or not. As of now this is possible, because running the command

  build -f uuid-build.xml

we automatically change the mode attribute on the status element to 'deleted', if the last event element have a flag with the value of 'deleted'.

The reason for this modification is the simpler XPath expression.